Trabalho04/requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
Flask==2.2.2
Flask-Cors==3.0.10
grpcio==1.50.0
grpcio-tools==1.50.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pika==1.3.1
protobuf==4.21.9
requests==2.28.1
six==1.16.0
urllib3==1.26.12
Werkzeug==2.2.2
